JuriQuest/backend/requirements.txt
Flask
requests
beautifulsoup4
google-genai
python-dotenv